add hello podman image and point hello-world to it
The quay.io/podman/hello image should now be available for all arches.[1]
Lets add it to the shortnames.conf.

Also change the hello-world to point to it instead of the docker.io image.
shortnames.conf is only used in podman and other related containers
projects but not docker so it makes more sense to point users to the
podman image IMO.

[1] https://github.com/containers/podman/issues/13364

Signed-off-by: Paul Holzinger <pholzing@redhat.com>